+++ packages/react-universal/src/Context.tsx
@@ -7,16 +7,19 @@ const key: Emitted = {
   navigation: null
 };
 
-const emitted_context = React.createContext<Emitted>(key);
-const { Provider: EmittedProvider, Consumer: Curious } = emitted_context;
+const response_context = React.createContext<Emitted>(key);
+const {
+  Provider: ResponseProvider,
+  Consumer: ResponseConsumer
+} = response_context;
 
 const router_context = React.createContext<CuriRouter>(null);
 const { Provider: RouterProvider, Consumer: RouterConsumer } = router_context;
 
 export {
-  EmittedProvider,
-  Curious,
-  emitted_context,
+  ResponseProvider,
+  ResponseConsumer,
+  response_context,
   RouterProvider,
   RouterConsumer,
   router_context