+++ website/src/pages/Packages/react-dom/v2/contents.js
@@ -8,7 +8,7 @@ import { meta as useNavigationFocusMeta } from "./api/useNavigationFocus";
 import { meta as useNavigatingMeta } from "./api/useNavigating";
 import { meta as useLocationMeta } from "./api/useLocation";
 import { meta as useHrefMeta } from "./api/useHref";
-import { meta as CuriousMeta } from "./api/curious";
+import { meta as ResponseConsumerMeta } from "./api/responseconsumer";
 
 export default [
   {
@@ -33,7 +33,7 @@ export default [
       useNavigatingMeta,
       useLocationMeta,
       useHrefMeta,
-      CuriousMeta
+      ResponseConsumerMeta
     ]
   }
 ];